+++ app/lib/communicator/communicator.js
@@ -20,7 +20,7 @@ function getXFormInfo( survey ) {
     }
 
     return _request( {
-        url: getFormListUrl( survey.openRosaServer ),
+        url: getFormListUrl( survey.openRosaServer, survey.openRosaId ),
         auth: survey.credentials,
         headers: {
             cookie: survey.cookie
@@ -110,7 +110,7 @@ function getMaxSize( survey ) {
 
 function authenticate( survey ) {
     var options = {
-        url: getFormListUrl( survey.openRosaServer ),
+        url: getFormListUrl( survey.openRosaServer, survey.openRosaId ),
         auth: survey.credentials,
         // Formhub has a bug and cannot use the correct HEAD method.
         method: config[ 'linked form and data server' ][ 'legacy formhub' ] ? 'get' : 'head'
@@ -160,8 +160,10 @@ function getAuthHeader( url, credentials ) {
     } );
 }
 
-function getFormListUrl( server ) {
-    return ( server.lastIndexOf( '/' ) === server.length - 1 ) ? server + 'formList' : server + '/formList';
+function getFormListUrl( server, id ) {
+    var query = ( id ) ? '?formID=' + id : '';
+    var path = ( server.lastIndexOf( '/' ) === server.length - 1 ) ? 'formList' : '/formList';
+    return server + path + query;
 }
 
 function getSubmissionUrl( server ) {
@@ -269,8 +271,6 @@ function _findFormAddInfo( formListXml, survey ) {
     var found;
     var index;
     var error;
-
-    debug( 'looking for form object with id "' + survey.openRosaId + '" in formlist' );
 
     return new Promise( function( resolve, reject ) {
         // first convert to JSON to make it easier to work with